requirements.txt
asgiref==3.4.1
autopep8==1.5.7
cffi==1.14.6
cryptography==3.4.7
Django==3.2.5
django-cors-headers==3.7.0
djangorestframework==3.12.4
Jinja2==3.0.1
MarkupSafe==2.0.1
mysqlclient==2.0.3
pycodestyle==2.7.0
pycparser==2.20
PyMySQL==1.0.2
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
